chore(mcp): remove support for extension protocol v1
All clients use protocol v2 by now. Remove the v1 handlers on both the
relay and the extension side and collapse the abstractions that only
existed to support two protocol versions:

- Relay: drop ExtensionProtocolV1 and the ExtensionProtocolHandler
  interface, always use ExtensionProtocolV2.
- Extension: fold the v2 protocol handler into RelayConnection, drop
  the eager (v1) pending connection path and the protocolVersion
  plumbing. The connect page now rejects clients requesting protocol
  version < 2.
