+++ packages/playwright-core/src/tools/mcp/cdpRelay.ts
@@ -21,9 +21,8 @@
  * - /cdp/guid - Full CDP interface for Playwright MCP
  * - /extension/guid - Extension connection
  *
- * Protocol version is controlled by PLAYWRIGHT_EXTENSION_PROTOCOL env variable:
- * - v1: single-tab, extension manages debugger attachment
- * - v2 (default): multi-tab, relay manages debugger via chrome.* APIs
+ * The protocol version advertised to the extension can be overridden with the
+ * PLAYWRIGHT_EXTENSION_PROTOCOL env variable (used in tests).
  */
 
 import { spawn } from 'child_process';
@@ -38,13 +37,12 @@ import { registry } from '../../server/registry/index';
 import { playwrightExtensionId } from '../utils/extension';
 import { addressToString } from '../utils/mcp/http';
 import { logUnhandledError } from './log';
-import { ExtensionProtocolV1 } from './cdpRelayV1';
 import { ExtensionProtocolV2 } from './cdpRelayV2';
 import * as protocol from './protocol';
 
 import type websocket from 'ws';
-import type { ExtensionCommand, ExtensionEvents } from './protocol';
-import type { CDPMessage, ExtensionProtocolHandler } from './cdpRelayHandler';
+import type { ExtensionCommandV2, ExtensionEventsV2 } from './protocol';
+import type { CDPMessage } from './browserModel';
 import type { WebSocket, WebSocketServer } from 'ws';
 
 
@@ -69,25 +67,21 @@ export class CDPRelayServer {
   private _cdpConnection: WebSocket | null = null;
   private _extensionConnection: ExtensionConnection | null = null;
   private _protocolVersion: number;
-  private _handler: ExtensionProtocolHandler;
+  private _handler: ExtensionProtocolV2;
   private _extensionConnectionPromise = new ManualPromise<void>();
 
   constructor(server: http.Server, browserChannel: string, executablePath?: string) {
     this._wsHost = addressToString(server.address(), { protocol: 'ws' });
     this._browserChannel = browserChannel;
     this._executablePath = executablePath;
-    this._protocolVersion = parseInt(process.env.PLAYWRIGHT_EXTENSION_PROTOCOL ?? protocol.DEFAULT_VERSION.toString(), 10);
+    this._protocolVersion = parseInt(process.env.PLAYWRIGHT_EXTENSION_PROTOCOL ?? protocol.VERSION.toString(), 10);
 
     const sendCommand = (method: string, params: any): Promise<any> => {
       if (!this._extensionConnection)
         throw new Error('Extension not connected');
-      return this._extensionConnection.send(method as keyof ExtensionCommand, params);
+      return this._extensionConnection.send(method as keyof ExtensionCommandV2, params);
     };
-
-    if (this._protocolVersion >= 2)
-      this._handler = new ExtensionProtocolV2(sendCommand);
-    else
-      this._handler = new ExtensionProtocolV1(sendCommand);
+    this._handler = new ExtensionProtocolV2(sendCommand);
 
     const uuid = crypto.randomUUID();
     this._cdpPath = `/cdp/${uuid}`;
@@ -290,7 +284,7 @@ class ExtensionConnection {
   private readonly _callbacks = new Map<number, { resolve: (o: any) => void, reject: (e: Error) => void, error: Error }>();
   private _lastId = 0;
 
-  onmessage?: <M extends keyof ExtensionEvents>(method: M, params: ExtensionEvents[M]['params']) => void;
+  onmessage?: <M extends keyof ExtensionEventsV2>(method: M, params: ExtensionEventsV2[M]['params']) => void;
   onclose?: (reason: string) => void;
 
   constructor(ws: WebSocket) {
@@ -300,7 +294,7 @@ class ExtensionConnection {
     this._ws.on('error', this._onError.bind(this));
   }
 
-  async send<M extends keyof ExtensionCommand>(method: M, params: ExtensionCommand[M]['params']): Promise<any> {
+  async send<M extends keyof ExtensionCommandV2>(method: M, params: ExtensionCommandV2[M]['params']): Promise<any> {
     if (this._ws.readyState !== ws.OPEN)
       throw new Error(`Unexpected WebSocket state: ${this._ws.readyState}`);
     const id = ++this._lastId;
@@ -349,7 +343,7 @@ class ExtensionConnection {
     } else if (object.id) {
       debugLogger('← Extension: unexpected response', object);
     } else {
-      this.onmessage?.(object.method! as keyof ExtensionEvents, object.params);
+      this.onmessage?.(object.method! as keyof ExtensionEventsV2, object.params);
     }
   }
