+++ tests/extension/playwright.config.ts
@@ -17,9 +17,8 @@
 import { defineConfig } from '@playwright/test';
 
 import type { TestOptions } from '../mcp/fixtures';
-import type { ExtensionTestOptions } from './extension-fixtures';
 
-export default defineConfig<TestOptions & ExtensionTestOptions>({
+export default defineConfig<TestOptions>({
   testDir: './',
   fullyParallel: true,
   forbidOnly: !!process.env.CI,
@@ -30,7 +29,6 @@ export default defineConfig<TestOptions & ExtensionTestOptions>({
     ['../config/parquetReporter.ts'],
   ] : 'list',
   projects: [
-    { name: 'chromium', use: { mcpBrowser: 'chromium', protocolVersion: 2 } },
-    { name: 'chromium (legacy v1)', use: { mcpBrowser: 'chromium', protocolVersion: 1 } },
+    { name: 'chromium', use: { mcpBrowser: 'chromium' } },
   ],
 });
